Moving picviz into output
